refactor file structure for fh-v8
